chore: refactor to use TableVariableItem
