+++ packages/frontend/src/components/FlowStepTestController/index.tsx
@@ -10,7 +10,6 @@ import {
   HStack,
   Text,
   Tooltip,
-  useDisclosure,
   VStack,
 } from '@chakra-ui/react'
 import { Button, Infobox } from '@opengovsg/design-system-react'
@@ -21,11 +20,9 @@ import { useStepMetadata } from '@/hooks/useStepMetadata'
 
 import { EDITOR_MARGIN_TOP_NUM } from '../Editor/constants'
 import ErrorResult from '../ErrorResult'
-import { isMultiRowStep } from '../MultiRowResultVariables/utils'
 import WebhookUrlInfo from '../WebhookUrlInfo'
 
 import { flowStepTestControllerStyles } from './styles'
-import TestMultiRowResultModal from './TestMultiRowResultModal'
 import TestResult from './TestResult'
 import { useTestDetails } from './useTestDetails'
 import { getInfoBoxDetails, matchParamsToDataIn } from './utils'
@@ -99,11 +96,6 @@ export default function FlowStepTestController(
     testExecutionSteps,
     varInfoMap,
   } = useContext(EditorContext)
-  const {
-    isOpen: isModalOpen,
-    onOpen: onModalOpen,
-    onClose: onModalClose,
-  } = useDisclosure()
   const formContext = useFormContext()
 
   const { isIfThenStep, isTrigger, selectedActionOrTrigger, substeps } =
@@ -361,7 +353,6 @@ export default function FlowStepTestController(
               isMock={currentTestExecutionStep?.metadata?.isMock}
               isOpen={isTestResultOpen}
               isIfThenStep={isIfThenStep}
-              onModalOpen={onModalOpen}
             />
           </VStack>
         ) : (
@@ -400,13 +391,6 @@ export default function FlowStepTestController(
           </VStack>
         )}
       </VStack>
-      {isMultiRowStep(step) && (
-        <TestMultiRowResultModal
-          isOpen={isModalOpen}
-          onClose={onModalClose}
-          currentExecutionStep={currentTestExecutionStep}
-        />
-      )}
     </>
   )
 }
