--- packages/frontend/src/components/MultiRowResultVariables/index.tsx
@@ -1,95 +1,0 @@
-import { IAction, IStep, ITrigger } from '@plumber/types'
-
-import { useMemo } from 'react'
-import { MdOpenInNew } from 'react-icons/md'
-import { Box } from '@chakra-ui/react'
-
-import { Variable } from '@/helpers/variables'
-
-import { VariableItem } from '../VariablesList'
-
-interface TestMultiRowResultProps {
-  step: IStep
-  selectedActionOrTrigger: ITrigger | IAction | undefined
-  variables: Variable[] | null
-  isMock?: boolean
-  onModalOpen?: () => void
-}
-
-export default function MultiRowResultVariables(
-  props: TestMultiRowResultProps,
-): JSX.Element {
-  const { step, variables, onModalOpen } = props
-
-  const isDataVar = (v: Variable) => v.name.split('.').pop() === 'data'
-  const isRowsFoundVar = (v: Variable) =>
-    v.name.split('.').pop() === 'rowsFound'
-
-  const { variablesWithModal, variableListVariables } = useMemo(() => {
-    if (!variables?.length) {
-      return { variablesWithModal: [], variableListVariables: [] }
-    }
-
-    const dataVariable = variables.find(isDataVar)
-    const rowsFoundVariable = variables.find(isRowsFoundVar)
-    const numRowsFound = rowsFoundVariable?.value || 0
-
-    // NOTE: do not expose the column variables as there is not real use case
-    // for users at this point.
-    // const columnVariables = dataVariable ? getColumnValues(dataVariable) : []
-    const rowsFoundVariables = variables.filter(isRowsFoundVar)
-
-    const variableListVariables = [...rowsFoundVariables]
-
-    if (numRowsFound === 0) {
-      return {
-        variablesWithModal: [],
-        variableListVariables: dataVariable
-          ? [dataVariable, ...variableListVariables]
-          : variableListVariables,
-      }
-    }
-
-    return {
-      variablesWithModal: dataVariable ? [dataVariable] : [],
-      variableListVariables,
-    }
-  }, [variables])
-
-  if (step.status !== 'completed') {
-    return <></>
-  }
-
-  // No data only happens if user hasn't executed yet, or step returned null.
-  if (!variablesWithModal && !variableListVariables) {
-    return <></>
-  }
-
-  return (
-    <Box
-      data-test="variable-suggestion-group"
-      maxH={64}
-      overflowY="auto"
-      p="1rem"
-    >
-      {variablesWithModal.map((variable) => (
-        <VariableItem
-          key={`variable-${variable.name}`}
-          variable={variable}
-          onClick={onModalOpen}
-          withIcon={MdOpenInNew}
-        />
-      ))}
-      {variableListVariables.length > 0 &&
-        variableListVariables
-          .filter((v): v is Variable => 'label' in v && 'type' in v)
-          .map((v, index) => (
-            <VariableItem
-              key={`variable-${v.id}-${v.name}`}
-              variable={v}
-              isLast={index === variableListVariables.length - 1}
-            />
-          ))}
-    </Box>
-  )
-}
